fix: resolve validation and async issues in simulation pipeline
- Fix Depends annotation conflict in simulation route (cannot use both Annotated and =Depends())
- Change conversation_history type from Dict[str, str] to Dict[str, Any] to allow int turn_number and Optional sentiment
- Remove incorrect await on non-async select_focus_concept() call
- Change SimulationTurnSchema.latency_ms from int to float (LLM returns float timestamps)

+++ src/api/schemas.py
@@ -322,7 +322,7 @@ class SimulationTurnSchema(BaseModel):
     persona_name: str
     strategy_selected: Optional[str] = None
     should_continue: bool = True
-    latency_ms: int = 0
+    latency_ms: float = 0.0
 
 
 class SimulationRequest(BaseModel):
